% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BY2011-datasets.R
\docType{data}
\name{BY2011_annual}
\alias{BY2011_annual}
\title{BY2011 annual emissions (all categories)}
\format{A data frame with 990,621 rows and 6 columns:
\describe{
  \item{year}{(int) four-digit year}
  \item{cnty_abbr}{(chr) county identifier: ALA, CC, MAR, NAP, SF, SM, SNC, SOL, SON}
  \item{cat_id}{(int) category identifier}
  \item{pol_abbr}{(chr) pollutant identifier: PM, PM10, PM2.5, TOG, ROG, NOx, SO2, CO, CO2, CO2_bio, CH4, N2O, HFC+PFC, SF6}
  \item{ems_qty}{(int) estimated emissions}
  \item{ems_unit}{(chr) ton/yr}
}}
\usage{
BY2011_annual
}
\description{
A dataset containing estimated emissions by year, county,
category, and pollutant. BY2011 means "Base Year 2011".
}
\note{
Please use this instead of \code{BY2011_data}. Thanks!
}
\examples{
library(inventory)
BY2011_annual \%>\% filter(cat_id \%in\% BY2011_sets[["Area sources"]])
BY2011_annual \%>\% filter(cat_id \%not_in\% BY2011_sets[["Unreported"]])

}
\seealso{
BY2011_sets BY2011_categories with_hierarchy sum_annual_emissions_by
}
\keyword{datasets}
