% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BY2011-datasets.R
\docType{data}
\name{BY2011_categories}
\alias{BY2011_categories}
\title{BY2011 category IDs and descriptors (nested headings)}
\format{A data frame with columns \code{cat_id}, \code{cat_h1}, \code{cat_h2}, ... \code{cat_h6}}
\usage{
BY2011_categories
}
\description{
A named list containing useful sets of BY2011 category identifiers (IDs).
To see them in the console, type `BY2011_sets` (no quotes), then hit Enter.
}
\examples{
library(inventory)
View(BY2011_categories)
BY2011_annual \%>\% with_hierarchy(BY2011_categories)

}
\seealso{
BY2011_annual BY2011_sets with_hierarchy
}
\keyword{datasets}
