% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BY2011-datasets.R
\docType{data}
\name{BY2011_data}
\alias{BY2011_data}
\title{DEPRECATED: BY2011 annual emissions (all categories)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 991892 rows and 6 columns.}
\usage{
BY2011_data
}
\description{
DEPRECATED: BY2011 annual emissions (all categories)
}
\note{
Deprecated in favor of \code{BY2011_annual}.
}
\seealso{
BY2011_annual
}
\keyword{datasets}
