% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BY2011-datasets.R
\docType{data}
\name{BY2011_GWPs}
\alias{BY2011_GWPs}
\title{Global Warming Potential (GWP) factors used for/in/with BY2011}
\format{A named vector, with pollutant abbreviations used as names.}
\usage{
BY2011_GWPs
}
\description{
Type \code{show(BY2011_GWPs)} to see the values.
}
\examples{
library(inventory)
show(BY2011_GWPs)

}
\seealso{
BY2011_annual as_MMTCO2eq
}
\keyword{datasets}
