% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BY2011-datasets.R
\docType{data}
\name{BY2011_sets}
\alias{BY2011_sets}
\title{Notable sets of BY2011 categories}
\format{A data frame with 990,621 rows and 6 columns:
\describe{
  \item{"Area sources"}{(int) area source category IDs}
  \item{"Point sources"}{(int) point source category IDs}
  \item{"Motor vehicles"}{(int) motor vehicle category IDs}
  \item{"Unreported"}{(int) Category IDs corresponding to ships > 3 nautical miles, as well as those corresponding to biogenics}
}}
\usage{
BY2011_sets
}
\description{
A named list containing useful sets of BY2011 category identifiers (IDs).
To see them in the console, type \code{show(BY2011_sets)} (no quotes), then hit Enter.
}
\examples{
library(inventory)
options(width = 80)
show(BY2011_sets)
BY2011_annual \%>\% filter(cat_id \%in\% BY2011_sets[["Area sources"]])
BY2011_annual \%>\% filter(cat_id \%not_in\% BY2011_sets[["Unreported"]])

}
\seealso{
BY2011_annual
}
\keyword{datasets}
