% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_point_source_category.R
\name{is_point_source_category}
\alias{is_point_source_category}
\title{is_point_source_category}
\usage{
is_point_source_category(ids)
}
\description{
is_point_source_category
}
\note{
NOT exported because BY2008 and BY2005 contain functions with the same
  name. Invoked within the `DataBank` package like
  `BY2011:::is_point_source_category()`.
}
